reorganize scripts vs. utils
