+++ .env.test.e2e
@@ -1,3 +1,4 @@
+export PW_EXPERIMENTAL_TS_ESM
 export PW_MODE=e2e
 VITE_TEMPORAL_PORT="7233"
 VITE_API="http://localhost:8233"
