+++ .env.test.integration
@@ -1,3 +1,4 @@
+export PW_EXPERIMENTAL_TS_ESM=1
 export PW_MODE=integration
 VITE_TEMPORAL_PORT="4444"
 VITE_API="http://localhost:5555"
