+++ scripts/start-temporal-server.ts
@@ -1,107 +1,25 @@
-import { join } from 'path';
-import waitForPort from 'wait-port';
-import { $, chalk } from 'zx';
+import {
+  type TemporalServer,
+  type TemporalServerOptions,
+  createTemporalServer,
+} from '../temporal/temporal-server';
 
-export type TemporalServer = {
-  shutdown: () => Promise<number | null>;
-  ready: () => Promise<boolean>;
+import yargs from 'yargs/yargs';
+
+const args = yargs(process.argv.slice(2)).parse();
+
+const options: TemporalServerOptions = {
+  port: args['port'],
+  uiPort: args['uiPort'] ?? args['ui-port'],
+  path: args['path'],
+  logLevel: args['logLevel'] ?? args['log-level'],
+  codecEndpoint: args['codecEndpoint'] ?? args['codec-endpoint'],
 };
 
-const localCLIPath = join(process.cwd(), 'bin', 'cli', 'temporal');
+const server: TemporalServer = await createTemporalServer(options);
 
-type TemporalServerOptions = {
-  port?: number;
-  uiPort?: number;
-  path?: string;
-  logLevel?: string;
-  codecEndpoint?: string;
-};
+await server.ready();
 
-const warn = (message: Parameters<typeof console.warn>[0]) => {
-  console.warn(`${chalk.bgYellow.black('WARN')}: ${message}`);
-};
-
-const getCLIPath = async (cliPath = localCLIPath): Promise<string | void> => {
-  const stylizedPath = chalk.yellowBright(cliPath);
-
-  console.log(chalk.yellow(`Checking Temporal CLI at ${stylizedPath}…`));
-
-  const { stdout, exitCode } = await $`${cliPath} -v`.quiet().nothrow();
-
-  if (exitCode === 0) {
-    console.log(
-      chalk.greenBright(
-        `Temporal CLI found at ${stylizedPath}:\n\t`,
-        '→',
-        chalk.green(stdout.trim()),
-      ),
-    );
-
-    return cliPath;
-  }
-
-  const { stdout: globalPath } = await $`which temporal`.nothrow();
-
-  if (globalPath && cliPath !== globalPath.trim())
-    return getCLIPath(globalPath.trim());
-
-  warn("Couldn't find Temporal CLI. Skipping…");
-};
-
-export const createTemporalServer = async ({
-  port = 7233,
-  uiPort = port + 1000,
-  path = localCLIPath,
-  logLevel = 'fatal',
-  codecEndpoint,
-}: TemporalServerOptions = {}) => {
-  const cliPath = await getCLIPath(path);
-
-  const flags = [
-    `--port=${port}`,
-    `--ui-port=${uiPort}`,
-    `--log-level=${logLevel}`,
-  ];
-
-  if (codecEndpoint) {
-    flags.push(`--ui-codec-endpoint=${codecEndpoint}`);
-  }
-
-  const temporal = $`${cliPath} server start-dev ${flags}`.quiet();
-
-  temporal.catch(async ({ stdout, stderr, exitCode }) => {
-    if (exitCode) {
-      try {
-        const { error }: { error: string } = JSON.parse(stdout);
-
-        if (error.includes('address already in use')) {
-          return warn(
-            `Port ${port} is already in use. Falling back to whatever is running on that port.`,
-          );
-        }
-
-        throw new Error(stderr ?? stdout);
-      } catch (error) {
-        throw new Error(stderr ?? stdout);
-      }
-    }
-  });
-
-  const shutdown = async () => {
-    await temporal.kill();
-    return await temporal.exitCode;
-  };
-
-  const ready = async () => {
-    const ports = await Promise.all([
-      waitForPort({ port, output: 'silent' }),
-      waitForPort({ port: uiPort, output: 'silent' }),
-    ]);
-    return ports.every(({ open }) => open);
-  };
-
-  return {
-    ready,
-    shutdown,
-  };
-};
+process.on('beforeExit', async () => {
+  await server.shutdown();
+});
