--- temporal-fixtures/activities/index.ts
@@ -1,2 +1,0 @@
-export { default as echo } from '$temporal-fixtures/activities/echo';
-export { default as sideEffect } from '$temporal-fixtures/activities/side-effect.js';
