+++ temporal/activities/index.ts
@@ -1,0 +1,2 @@
+export { default as echo } from './echo';
+export { default as sideEffect } from './side-effect';
