+++ tests/integration/workflows-list-time-frames.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect } from '@playwright/test';
-import { mockDate } from '~/utilities/mock-date.js';
-import { mockClusterApi } from '~/utilities/mock-apis.js';
-import { setLocalStorage } from '~/utilities/mock-local-storage';
+import { mockDate } from '~test-utilities/mock-date';
+import { mockClusterApi } from '~/test-utilities/mock-apis';
+import { setLocalStorage } from '~/test-utilities/mock-local-storage';
 
 const workflowsUrl = '/namespaces/default/workflows';
 const workflowsApiMatch = new RegExp('/api/v1/namespaces/default/workflows');
