+++ tests/integration/workflows-list.spec.ts
@@ -1,6 +1,6 @@
 import { test, expect } from '@playwright/test';
-import { workflowsApi, mockClusterApi } from '~/utilities/mock-apis.js';
-import { setLocalStorage } from '~/utilities/mock-local-storage';
+import { workflowsApi, mockClusterApi } from '~/test-utilities/mock-apis';
+import { setLocalStorage } from '~/test-utilities/mock-local-storage';
 const workflowsUrl = '/namespaces/default/workflows';
 
 test.beforeEach(async ({ page }) => {
